fix: small collection of bugfixed debugger hints and warnings (#801)
* small collection of bugfixed debugger hints / warnings

* 'toStdVector' has been explicitly marked deprecated

* fixed double zip naming in artifacts

* V4L2 WebUI Fix

* Some code fixes based on alerts from lgtm.com

* only execute dynamic v4l2 enum code, if V4L2_AVAIL

* very high critical bugfix ;)

* merge fix

* some lgtm.com fixes

* lgtm fixes

* undo localtime_r fix

+++ include/utils/Image.h
@@ -203,7 +203,7 @@ public:
 		assert(other._width == _width);
 		assert(other._height == _height);
 
-		memcpy(_pixels, other._pixels, _width*_height*sizeof(Pixel_T));
+		memcpy(_pixels, other._pixels, _width * _height * sizeof(Pixel_T));
 	}
 
 	///
@@ -258,7 +258,7 @@ public:
 		_height = 1;
 		_pixels = new Pixel_T[2];
 		_endOfPixels = _pixels + 1;
-		memset(_pixels, 0, _width * _height * sizeof(Pixel_T));
+		memset(_pixels, 0, (unsigned long) _width * _height * sizeof(Pixel_T));
 	}
 
 private:
